+++ CHANGELOG.md
@@ -3,12 +3,12 @@
 ## [Unreleased]
 
 ### Added
-
 - Add principal variation search
 - Add internal iterative deepening
 - Add late move reduction
 - Add killer heuristic
 - Add basic null move pruning
+- Add basic futility pruning
 - Add basic mobility evaluation
 - Add static exchange evaluation